Rework the entanglement of the URI, Playlist and Song components.
Rename SongData module to Song.
Variants will be phased out. 'irealb' (irealpro) is standard.
However, 'irealbook' as input will be  supported as usual.
Deprecate the export methods of these components in favour of
the much cleaner 'as_string' methods.
Enhance the irealcvt program with HTML splitting and listing,
and automatically combining of multiple input files.
Fix a problem with the CGI backend and playlists.
Package App::Packager for convenience.